+++ src/realm/views.hpp
@@ -8,7 +8,7 @@
 
 namespace realm {
 
-const size_t detached_ref = size_t(-1);
+const int64_t detached_ref = -1;
 
 // This class is for common functionality of ListView and LinkView which inherit from it. Currently it only
 // supports sorting.
